+++ lib/govuk_content_item_loader/test_helpers.rb
@@ -21,6 +21,11 @@ module GovukConditionalContentItemLoaderTestHelpers
     stub_publishing_api_isnt_available
   end
 
+  def stub_conditional_content_loader_has_gone_item(base_path)
+    stub_content_store_has_gone_item(base_path)
+    stub_publishing_api_graphql_has_gone_item(base_path)
+  end
+
 private
 
   def content_item_for_base_path(base_path)