+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add `stub_conditional_content_loader_isnt_available` test helper.
+* Add `stub_conditional_content_loader_has_gone_item` test helper.
 
 ## 2.0.0